# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShadyCard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List products
      parameters:
        - in: query
          name: categoryId
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        categoryId: { type: string }
        categoryName: { type: string }
        image: { type: string }
        priceUsd: { type: number }
        priceSyp: { type: number }
        productType: { type: string, enum: [amount, package] }
        available: { type: boolean }
        minQty: { type: number }
        maxQty: { type: number }
        description: { type: string }
        featured: { type: boolean }
      required: [id, name, categoryId, categoryName, image, priceUsd, priceSyp, productType, available, featured]
...